Update on id attribute's name and Bug Fix on Save Products
The id is the barcode, but in some parts of the code it was still named id. So, I updated them and finally found the solution to the error that prevented the product from being saved in the database. I did not specify the name of the columns in ContentValues in the Java ProductDAO file. Now, duly specified, the recording of data in the db occurs normally.